requirements.txt
discord.py==2.3.2
wavelink==2.6.1
requests==2.31.0


